+++ CMakeLists.txt.user
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE QtCreatorProject>
-<!-- Written by QtCreator 3.5.1, 2016-10-16T03:45:20. -->
+<!-- Written by QtCreator 3.5.1, 2016-10-16T13:36:08. -->
 <qtcreator>
  <data>
   <variable>EnvironmentId</variable>
