+++ src/bullet.hpp
@@ -1,18 +1,20 @@
 #pragma once
 
-#include <gameentity.h>
-#include <point2d.hpp>
-#include <box2d.hpp>
+#include "gameentity.hpp"
+#include "movable.hpp"
+#include "point2d.hpp"
+#include "box2d.hpp"
 #include <cmath>
 #include <initializer_list>
 #include <functional>
 #include <ostream>
 
-class Movable : public GameEntity
+class Bullet : public Movable
 {
 public:
-  virtual ~Movable() = default;
-  virtual void Move();
+  Bullet() = default;
+
+  void Move() override;
 
 private:
   Point2D m_coordinate = {0.0f, 0.0f};
