+++ src/gameentity.cpp
@@ -1,7 +1,7 @@
 #pragma once
 
-#include <point2d.hpp>
-#include <box2d.hpp>
+#include "point2d.hpp"
+#include "box2d.hpp"
 #include <cmath>
 #include <initializer_list>
 #include <functional>
