+++ src/space.cpp
@@ -1,4 +1,4 @@
-#include "space.h"
+#include "space.hpp"
 
 //Space::Space()
 //{
