+++ src/space.hpp
@@ -1,19 +1,19 @@
 #pragma once
 
-#include <point2d.hpp>
-#include <box2d.hpp>
+#include "gameentity.hpp"
+#include "point2d.hpp"
+#include "box2d.hpp"
 #include <cmath>
 #include <initializer_list>
 #include <functional>
 #include <ostream>
 
-class GameEntity
+class Space : public GameEntity
 {
 public:
-  virtual ~GameEntity() = 0;
-  virtual toString() = 0;
+  Space() = default;
 
 private:
-  Point2D m_coordinate = {0.0f, 0.0f};
   Box2D m_body = {0.0f, 0.0f, 0.0f, 0.0f};
 };
+
